Added engine_import implementation.
