--- addons/source-python/docs/source-python/source/developing/modules/entities.engines.csgo.csgo.rst
@@ -1,7 +1,0 @@
-entities.engines.csgo.csgo module
-==================================
-
-.. automodule:: entities.engines.csgo.csgo
-    :members:
-    :undoc-members:
-    :show-inheritance:
