--- addons/source-python/docs/source-python/source/developing/modules/entities.engines.csgo.rst
@@ -1,10 +1,0 @@
-entities.engines.csgo package
-==============================
-
-Module contents
----------------
-
-.. automodule:: entities.engines.csgo
-    :members:
-    :undoc-members:
-    :show-inheritance:
