--- addons/source-python/docs/source-python/source/developing/modules/entities.engines.rst
@@ -1,18 +1,0 @@
-entities.engines package
-=========================
-
-Subpackages
------------
-
-.. toctree::
-    :titlesonly:
-
-    entities.engines.csgo
-
-Module contents
----------------
-
-.. automodule:: entities.engines
-    :members:
-    :undoc-members:
-    :show-inheritance:
