--- addons/source-python/docs/source-python/source/developing/modules/players._base.rst
@@ -1,12 +1,0 @@
-players._base module
-====================
-
-.. automodule:: players._base
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-.. autoclass:: _players.PlayerMixin
-    :members:
-    :undoc-members:
-    :show-inheritance:
