--- addons/source-python/docs/source-python/source/developing/modules/players.engines.bms.rst
@@ -1,10 +1,0 @@
-players.engines.bms package
-============================
-
-Module contents
----------------
-
-.. automodule:: players.engines.bms
-    :members:
-    :undoc-members:
-    :show-inheritance:
