--- addons/source-python/docs/source-python/source/developing/modules/players.engines.csgo.rst
@@ -1,10 +1,0 @@
-players.engines.csgo package
-=============================
-
-Module contents
----------------
-
-.. automodule:: players.engines.csgo
-    :members:
-    :undoc-members:
-    :show-inheritance:
