--- addons/source-python/docs/source-python/source/developing/modules/players.engines.l4d2.rst
@@ -1,10 +1,0 @@
-players.engines.l4d2 package
-=============================
-
-Module contents
----------------
-
-.. automodule:: players.engines.l4d2
-    :members:
-    :undoc-members:
-    :show-inheritance:
