--- addons/source-python/docs/source-python/source/developing/modules/players.engines.orangebox.cstrike.rst
@@ -1,7 +1,0 @@
-players.engines.orangebox.cstrike module
-=========================================
-
-.. automodule:: players.engines.orangebox.cstrike
-    :members:
-    :undoc-members:
-    :show-inheritance:
