--- addons/source-python/docs/source-python/source/developing/modules/players.engines.orangebox.rst
@@ -1,18 +1,0 @@
-players.engines.orangebox package
-==================================
-
-Submodules
-----------
-
-.. toctree::
-   :titlesonly:
-
-   players.engines.orangebox.cstrike
-
-Module contents
----------------
-
-.. automodule:: players.engines.orangebox
-    :members:
-    :undoc-members:
-    :show-inheritance:
