--- addons/source-python/docs/source-python/source/developing/modules/players.engines.rst
@@ -1,19 +1,0 @@
-players.engines package
-========================
-
-Subpackages
------------
-
-.. toctree::
-    :titlesonly:
-
-    players.engines.csgo
-    players.engines.orangebox
-
-Module contents
----------------
-
-.. automodule:: players.engines
-    :members:
-    :undoc-members:
-    :show-inheritance:
