--- addons/source-python/docs/source-python/source/developing/modules/weapons.engines.csgo.csgo.rst
@@ -1,7 +1,0 @@
-weapons.engines.csgo.csgo module
-=================================
-
-.. automodule:: weapons.engines.csgo.csgo
-    :members:
-    :undoc-members:
-    :show-inheritance:
