--- addons/source-python/docs/source-python/source/developing/modules/weapons.engines.csgo.rst
@@ -1,10 +1,0 @@
-weapons.engines.csgo package
-=============================
-
-Module contents
----------------
-
-.. automodule:: weapons.engines.csgo
-    :members:
-    :undoc-members:
-    :show-inheritance:
