--- addons/source-python/docs/source-python/source/developing/modules/weapons.engines.rst
@@ -1,18 +1,0 @@
-weapons.engines package
-========================
-
-Subpackages
------------
-
-.. toctree::
-    :titlesonly:
-
-    weapons.engines.csgo
-
-Module contents
----------------
-
-.. automodule:: weapons.engines
-    :members:
-    :undoc-members:
-    :show-inheritance:
