--- addons/source-python/packages/source-python/entities/engines/__init__.py
@@ -1,3 +1,0 @@
-# ../entities/engines/__init__.py
-
-"""Provides engine specific Entity based functionality."""
