--- addons/source-python/packages/source-python/entities/engines/csgo/__init__.py
@@ -1,16 +1,0 @@
-# ../entities/engines/csgo/__init__.py
-
-"""Provides CS:GO engine specific Entity based functionality."""
-
-# =============================================================================
-# >> IMPORTS
-# =============================================================================
-# Source.Python
-from entities._base import Entity as _Entity
-
-
-# =============================================================================
-# >> CLASSES
-# =============================================================================
-class Entity(_Entity):
-    """Class used to interact directly with entities."""
