+++ addons/source-python/packages/source-python/players/bms/entity.py
@@ -1,21 +1,11 @@
-# ../players/engines/bms/__init__.py
+# ../players/bms/entity.py
 
 """Provides BM:S specific Player based functionality."""
 
 # =============================================================================
-# >> IMPORTS
-# =============================================================================
-# Source.Python Imports
-#   Entities
-from entities.helpers import wrap_entity_mem_func
-#   Players
-from players._base import Player as _Player
-
-
-# =============================================================================
 # >> CLASSES
 # =============================================================================
-class Player(_Player):
+class Player(Player):
     """Class used to interact directly with players."""
 
     @wrap_entity_mem_func
