--- addons/source-python/packages/source-python/players/engines/__init__.py
@@ -1,3 +1,0 @@
-# ../players/engines/__init__.py
-
-"""Provides engine specific Player based functionality."""
