--- addons/source-python/packages/source-python/weapons/engines/__init__.py
@@ -1,3 +1,0 @@
-# ../weapons/engines/__init__.py
-
-"""Provides engine specific Weapon based functionality."""
