--- addons/source-python/packages/source-python/weapons/engines/csgo/__init__.py
@@ -1,16 +1,0 @@
-# ../weapons/engines/csgo/__init__.py
-
-"""Provides CS:GO engine specific weapon functionality."""
-
-# =============================================================================
-# >> IMPORTS
-# =============================================================================
-# Source.Python
-from weapons._base import Weapon as _Weapon
-
-
-# =============================================================================
-# >> CLASSES
-# =============================================================================
-class Weapon(_Weapon):
-    """Allows easy usage of the weapon's attributes."""
